Organize bookmarks by extracting components
